requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
kafka-python==2.0.2
